Changed int to double

+++ src/main/java/org/example/branch_java_demo/utils/CustomMethods.java
@@ -2,7 +2,7 @@ package org.example.branch_java_demo.utils;
 
 public class CustomMethods {
 
-    public int add(int x, int y) {
+    public double add(double x, double y) {
 
         return x + y;
     }
